Major: For "ScramblePanel.java" I've separated the Cube/Pyraminx/Megaminx into their own classes. We basically need to pass it a "String scramble" and it passes back a "BufferedImage".
It increases the Jar size by 7Kb so it's pretty major, but allows for some incredibly nice code clean up in a lot of places.

Separated out a "Polygon makeFace()" method in each new class so that we can create the face points in a separate place and pass the polygon into the "drawFace()" methods. This will free us to do more elaborate things with those Polygons later.